src/requirements.txt
flasgger==0.9.5
Flask==2.2.2
flask-jwt-extended==4.4.4
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
flask_migrate
gevent==22.10.2
gevent==22.10.2
marshmallow-sqlalchemy==0.28.1
psycopg2-binary==2.9.5
pydantic==1.10.4
pytest==6.2.5
redis==4.1.4
SQLAlchemy==2.0.2
